requirements.txt
absl-py==0.12.0
albumentations==0.5.2
apturl==0.5.2
argon2-cffi==20.1.0
asn1crypto==0.24.0
astor==0.8.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
Brlapi==0.6.6
cached-property==1.5.2
cachetools==4.2.2
certifi==2018.1.18
cffi==1.14.5
chardet==3.0.4
click==6.7
colorama==0.3.7
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distro-info===0.18ubuntu0.18.04.1
easydict==1.9
entrypoints==0.3
gast==0.2.2
geffnet==1.0.0
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.37.1
h5py==3.2.1
httplib2==0.9.2
idna==2.6
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==4.0.1
ipykernel==5.5.3
ipython==7.23.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
linecache2==1.0.0
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.3.4
MarkupSafe==1.0
matplotlib==3.4.1
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
netifaces==0.10.4
networkx==2.5.1
nose==1.3.7
notebook==6.3.0
numpy==1.20.2
oauth==1.0.1
oauthlib==3.1.0
olefile==0.45.1
opencv-python==4.5.1.48
opencv-python-headless==4.5.1.48
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pbr==3.1.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.15.8
psutil==5.4.2
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.8.1
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.7
pyRFC3339==1.0
pyrsistent==0.17.3
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2018.3
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==3.12
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
reportlab==3.4.0
requests==2.25.1
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
resnest==0.0.5
rsa==4.7.2
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.6.3
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Send2Trash==1.5.0
Shapely==1.7.1
simplejson==3.13.2
six==1.11.0
system-service==0.3
systemd-python==234
tensorboard==2.0.2
tensorflow-estimator==2.0.1
tensorflow-gpu==2.0.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
testresources==2.0.0
threadpoolctl==2.1.0
tifffile==2021.4.8
torch==1.7.1
torchaudio==0.7.2
torchvision==0.8.2
tornado==6.1
tqdm==4.60.0
traceback2==1.4.0
traitlets==5.0.5
typing-extensions==3.7.4.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unittest2==1.1.0
urllib3==1.22
usb-creator==0.3.3
wadllib==1.3.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xkit==0.0.0
zipp==3.4.1
zope.interface==4.3.2
